♻️ move getGithubUser query closer to component needing it
