+++ frontend/src/hooks/usePaymentRequests.ts
@@ -1,13 +1,7 @@
 import { gql } from "@apollo/client";
-import { useState } from "react";
-import { useHasuraLazyQuery, useHasuraMutation, useHasuraSubscription } from "src/hooks/useHasuraQuery";
+import { useHasuraMutation, useHasuraSubscription } from "src/hooks/useHasuraQuery";
 import { Currency, HasuraUserRole, PaymentStatus } from "src/types";
-import {
-  GetGithubUserQuery,
-  OnNewPaymentRequestsSubscription,
-  GithubUserFragment,
-  PaymentRequestFragment,
-} from "src/__generated/graphql";
+import { OnNewPaymentRequestsSubscription, PaymentRequestFragment } from "src/__generated/graphql";
 
 type Params = {
   projectId: string;
@@ -15,36 +9,13 @@ type Params = {
 };
 
 export default function usePaymentRequests({ projectId, onNewPaymentRequested }: Params) {
-  const fetchAllGithubRecipients = async (query: OnNewPaymentRequestsSubscription) => {
-    const allRecipientIds = new Set(
-      query.projectsByPk?.budgets.map(b => b.paymentRequests.map(r => r.recipientId)).flat() || []
-    );
-
-    for (const id of allRecipientIds) {
-      await getGithubUser({ variables: { githubUserId: id } });
-    }
-  };
-
   const getPaymentRequestsQuery = useHasuraSubscription<OnNewPaymentRequestsSubscription>(
     PAYMENT_REQUESTS_FOR_PROJECT_SUBSCRIPTION,
     HasuraUserRole.RegisteredUser,
     {
       variables: { projectId },
-      onData: ({ data }) => data.data && fetchAllGithubRecipients(data.data),
     }
   );
-
-  const [githubRecipients, setGithubRecipients] = useState<Record<number, GithubUserFragment>>({});
-
-  const addGithubRecipient = (query: GetGithubUserQuery) =>
-    setGithubRecipients(recipients => ({
-      ...recipients,
-      [query.fetchUserDetailsById.id]: query.fetchUserDetailsById,
-    }));
-
-  const [getGithubUser] = useHasuraLazyQuery<GetGithubUserQuery>(GET_GITHUB_USER_QUERY, HasuraUserRole.RegisteredUser, {
-    onCompleted: addGithubRecipient,
-  });
 
   const [requestNewPayment] = useHasuraMutation(REQUEST_PAYMENT_MUTATION, HasuraUserRole.RegisteredUser, {
     variables: { projectId },
@@ -55,16 +26,11 @@ export default function usePaymentRequests({ projectId, onNewPaymentRequested }:
 
   const toPaymentRequest = (paymentRequest: PaymentRequestFragment) => {
     const paidAmount = paymentRequest.payments.reduce((total, payment) => total + payment.amount, 0);
-    const recipient = githubRecipients[paymentRequest.recipientId];
 
     return {
       id: paymentRequest.id,
       amount: { value: paymentRequest.amountInUsd, currency: Currency.USD },
-      recipient: recipient && {
-        id: recipient.id,
-        login: recipient.login,
-        avatarUrl: recipient.avatarUrl,
-      },
+      recipientId: paymentRequest.recipientId,
       reason: paymentRequest.reason?.work_items?.at(0),
       status: paidAmount === paymentRequest.amountInUsd ? PaymentStatus.ACCEPTED : PaymentStatus.WAITING_PAYMENT,
       requestedAt: paymentRequest.requestedAt,
@@ -125,23 +91,6 @@ export const PAYMENT_REQUESTS_FOR_PROJECT_SUBSCRIPTION = gql`
           ...PaymentRequest
         }
       }
-    }
-  }
-`;
-
-export const GITHUB_USER_FRAGMENT = gql`
-  fragment GithubUser on User {
-    id
-    login
-    avatarUrl
-  }
-`;
-
-export const GET_GITHUB_USER_QUERY = gql`
-  ${GITHUB_USER_FRAGMENT}
-  query GetGithubUser($githubUserId: Int!) {
-    fetchUserDetailsById(userId: $githubUserId) {
-      ...GithubUser
     }
   }
 `;
